Refactor application-round single page
Remove secondary Review component.
